backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
djoser==2.1.0
webcolors==1.11.1
psycopg2-binary==2.8.6
Pillow==10.1.0
django-filter==2.3.0
django-extra-fields==2.0.3
